Write XLUM-files

Create XLUM-files using xml2::read_xml

Usage

write_xlum(x, file, encode_base64 = FALSE, ...)

Arguments

x

xlum_list, list, xml2::xml_document (required) : input object

file

character (required): valid file path and file name with ending .xlum (will be add automatically if needed)

encode_base64

logical (with default): enable/disable base64 encoding for values in the <curve/> node. Please note that this parameter works only if the input is of class xlum_list

...

further arguments to be passed to xml2::write_xml

Value

Creates an XML-file with ending *.xlum

Function version

0.1.0

See also

Author

Sebastian Kreutzer, Institute of Geography, Heidelberg University, Heidelberg (Germany)

How to cite

Kreutzer, S., 2025. write_xlum(): Write XLUM-files. Function version 0.1.0. In: Kreutzer, S., 2025. xlum: Read, Write, and Convert XLUM Data. R package version 0.1.0. https://r-lum.github.io/xlum/

Examples

file <- system.file("extdata/xlum_prototype.xlum", package="xlum")
out_file <- tempfile()

xlum_data <- read_xlum(file)
write_xlum(xlum_data, out_file)