Read, Write, and Convert xlum Data

Support for the XLUM-file format for exchange and long-term preservation of luminescence data. Facilitated are the import and export of XLUM-files and the conversion from other ASCII and binary formats commonly used to store luminescence (dating) data.

Details

Support contact

https://github.com/R-Lum/xlum/discussions

Bug reporting

https://github.com/R-Lum/xlum/issues

Project source code repository

https://github.com/R-Lum/xlum

Package maintainer

Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany),
sebastian.kreutzer@uni-heidelberg.de

Funding

Between 2020–2022, the work of Sebastian Kreutzer as maintainer of the package has received funding from the European Union’s Horizon 2020 research and innovation programme under the Marie Skłodowska-Curie grant agreement No 844457 (CREDit), and the work was carried out at the Geography & Earth Sciences, Aberystwyth University (United Kingdom)

See also

Author

Maintainer: Sebastian Kreutzer sebastian.kreutzer@uni-heidelberg.de (ORCID)

Other contributors:

Christoph Burow (ORCID) [contributor]