Convert Freiberg Instruments XSYG files to XLUM

Wrapped function to quickly convert Freiberg Instruments XSYG-files to xlum (files)

convert_xsyg2xlum(file, out_file = NULL)

Arguments

file

character (required): file

out_file

character (optional): output file, if set a file output is created

Value

Depending on the setting. If out_file == NULL an xml2::xml_document object is returned, if file is set the function attempts to write an *.xlum file`.

Function version

0.1.0

See also

Author

Sebastian Kreutzer, Institute of Geography, Heidelberg University, Heidelberg (Germany)

How to cite

Kreutzer, S., 2022. convert_xsyg2xlum(): Convert Freiberg Instruments XSYG files to XLUM. Function version 0.1.0. In: Kreutzer, S., 2022. xlum: Read, Write, and Convert XLUM Data. R package version 0.1.0.

Examples

## load package example data
## for loading own data use
## file <- file.choose()
file <- system.file("extdata/XSYG_file.xsyg", package = "Luminescence")
convert_xsyg2xlum(file)
#> {xml_document}
#> <xlum lang="en" formatVersion="1.0" flavour="string" author="string" license="CC BY 4.0" doi="" xmlns:xlum="http://xlum.r-luminescence.org">
#> [1] <sample name="NA" mineral="NA" latitude="52.416944" longitude="4.066389"  ...